+++ tests/test_sftp.py
@@ -1,6 +1,6 @@
 from ssh2.session import Session
 from ssh2.sftp import SFTP
-from ssh2.sftp_handle import SFTPAttributes
+from ssh2.sftp_handle import SFTPAttributes, SFTPHandle
 
 from .base_test import SSH2TestCase
 
@@ -21,3 +21,9 @@ class SFTPTestCase(SSH2TestCase):
     def test_session(self):
         session = Session()
         self.assertIsInstance(session, Session)
+
+    def test_sftp_handle(self):
+        session = Session()
+        sftp = SFTP(session)
+        sftp_fh = SFTPHandle(sftp)
+        self.assertFalse(sftp_fh.closed)