+++ Changelog.rst
@@ -11,6 +11,8 @@ Changes
 * Added `ssh2.session.Session.flag` function for enabling/disabling session flags like compression support.
 * Added `ssh2.session.userauth_keyboardinteractive_callback` for authentication using Python callback function,
   for example for Oauth and other two-factor (2FA) or more factor authentication. Thanks @MattCatz .
+* `ssh2.sftp_handle.SFTPHandle.closed` is now a public property indicating whether `ssh2.sftp_handle.SFTPHandle.close`
+  was called on a `SFTPHandle` or not.
 
 
 1.1.2